requirements_custom_model.txt
# Azure Form Recognizer Custom Model Requirements
flask==2.3.2
azure-ai-formrecognizer==3.1.1
azure-core==1.28.0
werkzeug==2.3.6
python-dotenv==1.0.0
pillow==10.0.0

# Optional: For image preprocessing
opencv-python==4.8.0.76
numpy==1.24.3

# Development
pytest==7.4.0
black==23.7.0
flake8==6.0.0
